+++ src/lib/rocket.cc
@@ -48,7 +48,7 @@ struct rocket_internal {
         using descriptor_t        = uint8_t;
         using descriptor_endian_t = little_endian;
         using sliding_window_t    = sliding_window<rocket_adaptor>;
-        enum class edge_type : size_t {
+        enum class edge_type : uint8_t {
             invalid,
             terminator,
             symbolwise,
